Give a better error message if token contains newlines
I constructed a token as `OAuth2(read("token", String))`, but my
token file had a trailing newline. GitHub stops parsing headers
after an empty line, so depending on where the header ended up
in the list of headers, this was causing all sorts of strange errors.
Detect this situation and throw an intelligent error in GitHub.jl instead.